git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/branch/prod/web-ui/work/src@15945 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ jnetcap/impl/com/untangle/jnetcap/NetcapUDPSession.java
@@ -181,7 +181,7 @@ public class NetcapUDPSession extends NetcapSession
     private static native int    icmpMerge( long sessionPointer, int id, long srcAddr, long dstAddr, 
                                             byte intf );
 
-    private static native int    mailboxPointer( long sessionPointer, boolean ifClient );
+    private static native long    mailboxPointer( long sessionPointer, boolean ifClient );
     
     /* This is for sending the data associated with a netcap_pkt_t structure */
     private static native int  send( long packetPointer );
@@ -229,7 +229,7 @@ public class NetcapUDPSession extends NetcapSession
             return read( 0 );
         }
 
-        public int pointer()
+        public long pointer()
         {
             return NetcapUDPSession.mailboxPointer( pointer.value(), ifClient );
         }
