+++ jvector/impl/com/untangle/jvector/IncomingSocketQueue.java
@@ -308,7 +308,7 @@ public class IncomingSocketQueue extends Sink
         mvpollNotifyObservers( this.pointer, poll());
     }
 
-    private native int create();
-    private native void mvpollNotifyObservers( int pointer, int eventMask );
+    private native long create();
+    private native void mvpollNotifyObservers( long pointer, int eventMask );
 }
